requirements.txt
altgraph==0.17.4
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.1
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
requests==2.32.3
setuptools==75.3.0
urllib3==2.2.3
Werkzeug==3.0.6
py-sc-client==0.3.1
